Rename legenda to legend
