+++ packages/app/src/components-styled/legend.tsx
@@ -9,7 +9,7 @@ interface LegendProps {
   items: LegendItem[];
 }
 
-export function Legenda({ items }: LegendProps) {
+export function Legend({ items }: LegendProps) {
   return (
     <List>
       {items.map(({ label, color, shape = 'line' }, i) => (
